Naparwa uprawnien

See merge request uam-informatyka/WirtualnePrzedszkole_PRI!23

+++ backend/src/main/java/com/example/wirtualneprzedszkole/controller/message/MessageController.java
@@ -1,6 +1,7 @@
 package com.example.wirtualneprzedszkole.controller.message;
 
 import com.example.wirtualneprzedszkole.mapper.message.MessageMapper;
+import com.example.wirtualneprzedszkole.model.UserRole;
 import com.example.wirtualneprzedszkole.model.dao.User;
 import com.example.wirtualneprzedszkole.model.dao.message.Message;
 import com.example.wirtualneprzedszkole.model.dao.message.UserMessage;
@@ -20,7 +21,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @RestController
 @RequiredArgsConstructor
@@ -31,29 +34,51 @@ public class MessageController {
     private final UserManagementService userManagementService;
     private final ChildService childService;
 
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
     @GetMapping
     public List<MessageDto> getAllSentMessages() {
         Long userId = userService.getCurrentUser().getId();
         return MessageMapper.mapMessagesToMessagesDto(messageService.getAllSentMessages(userId));
     }
 
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
     @PostMapping
     public ResponseEntity<MessageDto> sendMessage(Authentication authentication, @RequestBody SendMessageDto sendMessageDto) {
 
         sendMessageDto.setAuthor(getCurrentUser(authentication));
-        List<User> users = userManagementService.getAllUserByName(sendMessageDto.getTo());
+        //List<User> users = userManagementService.getAllUserByName(sendMessageDto.getTo());
+        List<User> users = userManagementService.getAllUserByEmail(sendMessageDto.getTo());
 
         if (users.size() > 0) {
-            return new ResponseEntity<>(MessageMapper.mapMessageToMessageDto(assignUserMessageToMsg(messageService
+            User user = userService.getCurrentUser();
+            if (user.getRole().getAuthority().equals("ADMIN"))
+                return new ResponseEntity<>(MessageMapper.mapMessageToMessageDto(assignUserMessageToMsg(messageService
                     .sendMessage(MessageMapper.SendMessageDtoMapToMessage(sendMessageDto), users), users)), HttpStatus.OK);
+            else if (user.getRole().getAuthority().equals("TEACHER")) {
+                Set<User> usersAllowed = usersAllowedForTeacher(user);
+                List<User> usersAllowedList = new ArrayList<>(usersAllowed);
+                users.retainAll(usersAllowedList);
+                if (users.size() > 0)
+                    return new ResponseEntity<>(MessageMapper.mapMessageToMessageDto(assignUserMessageToMsg(messageService
+                            .sendMessage(MessageMapper.SendMessageDtoMapToMessage(sendMessageDto), users), users)), HttpStatus.OK);
+                else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            else {
+                Set<User> usersAllowed = new HashSet<>(userManagementService.getAllByRole(UserRole.ADMIN));
+                usersAllowed.addAll(userManagementService.getAllByRole(UserRole.TEACHER));
+                List<User> usersAllowedList = new ArrayList<>(usersAllowed);
+                users.retainAll(usersAllowedList);
+                if (users.size() > 0)
+                    return new ResponseEntity<>(MessageMapper.mapMessageToMessageDto(assignUserMessageToMsg(messageService
+                            .sendMessage(MessageMapper.SendMessageDtoMapToMessage(sendMessageDto), users), users)), HttpStatus.OK);
+                else return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
         } else {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
     }
 
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN')")
     @PostMapping("to_parents")
     public MessageDto sendMessageToAllParents(Authentication authentication, @RequestBody SendMessageDto sendMessageDto) {
 
@@ -65,14 +90,18 @@ public class MessageController {
 
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER')")
     @PostMapping("to_class/{classId}")
-    public MessageDto sendMessageToClass(Authentication authentication,@PathVariable Long classId, @RequestBody SendMessageDto sendMessageDto) {
-
+    public ResponseEntity<MessageDto> sendMessageToClass(Authentication authentication,@PathVariable Long classId, @RequestBody SendMessageDto sendMessageDto) {
+        User user = userService.getCurrentUser();
         sendMessageDto.setAuthor(getCurrentUser(authentication));
-        List<Long> childrenIds = childService.getChildByClassIn(classId);
-        List<User> users = List.copyOf(userManagementService.getAllParentsFromClass(childrenIds));
-
-        return MessageMapper.mapMessageToMessageDto(assignUserMessageToMsg(messageService
-                .sendMessage(MessageMapper.SendMessageDtoMapToMessage(sendMessageDto), users), users));
+        if (user.getRole().getAuthority().equals("ADMIN")) {
+            return new ResponseEntity<>(send(classId, sendMessageDto), HttpStatus.OK);
+        }
+        else {
+            if (user.getClasses().stream().anyMatch(aClass -> aClass.getId().equals(classId))) {
+                return new ResponseEntity<>(send(classId, sendMessageDto), HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
@@ -91,14 +120,16 @@ public class MessageController {
         return MessageMapper.mapMessagesToMessageToRecipientsDto(messageService.getReceivedMessages(userId));
     }
 
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
     @GetMapping("sent_msg/{msgId}")
-    public MessageDtoWithFieldIsRead getSentMsg(@PathVariable Long msgId) {
+    public ResponseEntity<MessageDtoWithFieldIsRead> getSentMsg(@PathVariable Long msgId) {
         Message message = messageService.getSentMsg(msgId);
-        return MessageMapper.messageMapToMsgDtoWithFieldIsRead(message);
+        if (message.getAuthor().equals(userService.getCurrentUser()))
+            return new ResponseEntity<>(MessageMapper.messageMapToMsgDtoWithFieldIsRead(message), HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
-    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER')")
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
     @GetMapping("{msgId}/to/{userId}")
     public boolean checkMessageIsRead(@PathVariable Long msgId, @PathVariable Long userId) {
         return messageService.checkMessageIsRead(msgId, userId);
@@ -118,14 +149,35 @@ public class MessageController {
     }
 
 
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
     @DeleteMapping("{id}")
-    public void deleteClass(@PathVariable Long id) {
-        messageService.deleteMessage(id);
+    public ResponseEntity<Void> deleteClass(@PathVariable Long id) {
+        if (messageService.getSentMsg(id).getAuthor().equals(userService.getCurrentUser())) {
+            messageService.deleteMessage(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
+    @PreAuthorize("hasAnyRole('ROLE_ADMIN', 'ROLE_TEACHER', 'ROLE_PARENT')")
     @PatchMapping("/deleteReceivedMsg/{msgId}")
     public void deleteReceivedMsg(@PathVariable Long msgId) {
         messageService.deleteReceivedMsg(msgId, userService.getCurrentUser().getId());
     }
 
+    private MessageDto send(Long classId, SendMessageDto sendMessageDto) {
+        List<Long> childrenIds = childService.getChildByClassIn(classId);
+        List<User> users = List.copyOf(userManagementService.getAllParentsFromClass(childrenIds));
+
+        return MessageMapper.mapMessageToMessageDto(assignUserMessageToMsg(messageService
+                .sendMessage(MessageMapper.SendMessageDtoMapToMessage(sendMessageDto), users), users));
+    }
+
+    private Set<User> usersAllowedForTeacher(User user) {
+        Set<User> users = new HashSet<>();
+        user.getClasses().forEach(aClass -> aClass.getChildren().forEach(child -> users.addAll(child.getParents())));
+        users.addAll(userManagementService.getAllByRole(UserRole.TEACHER));
+        users.addAll(userManagementService.getAllByRole(UserRole.ADMIN));
+        return users;
+    }
 }
